Added server-side handling of file validation errors
ACF will throw errors during uploads but doesn’t make it easy to catch them. This change will intercept any upload errors and stop the submission if any exist.

This is mostly meant as a fallback and is not a user-friendly way of handling these errors. File validation should be implemented client-side in the future like all other validations.