Redesign automatic redirect in preparation for AJAX
Previously, the "redirect" argument defaulted to the current URL to
enable the automatic redirect back to the form after submission. With
AJAX, we need to separate automatic and manual redirects. This commit
changes the default "redirect" value to NULL and saves the current URL
as a hidden field. The NULL value will trigger an automatic redirect
back to the current URL.