Add support for arbitrary async submission steps on the client-side
Using reCAPTCHA together with AJAX is currently broken as they both hook
into and stop the form submission event. With this commit we introduce a
new concept of submission steps. Steps can be added with a priority and
during submission they will be executed in order. The steps are async
and need to call a callback for the submission to continue.